begin njt refactoring
