+++ malariagen_data/anoph/dipclust_params.py
@@ -1,6 +1,6 @@
 """Parameters for diplotype clustering functions."""
 
-from .diplotype_distance_params import distance_metric
+from .distance_params import distance_metric
 from .clustering_params import linkage_method
